complete overhaul for V9 push, long/int changes, list of arrays implemented; cast from long to int kept for many cases

+++ src/juicebox/Context.java
@@ -62,9 +62,9 @@ public class Context {
         this.zoom = zoom;
     }
 
-    public int getChrLength() {
-        return chromosome.getLength();
-    }
+    public long getChrLength() {
+		return chromosome.getLength();
+	}
 
     public Chromosome getChromosome() {
         return chromosome;
