+++ src/juicebox/data/censoring/CustomMZDRegionHandler.java
@@ -33,18 +33,18 @@ import org.broad.igv.util.Pair;
 import java.util.*;
 
 public class CustomMZDRegionHandler {
-
+    
     private final Map<String, Pair<List<MotifAnchor>, List<MotifAnchor>>> allRegionsForChr = new HashMap<>();
-    private final List<Integer> boundariesOfCustomChromosomeX = new ArrayList<>();
-    private final List<Integer> boundariesOfCustomChromosomeY = new ArrayList<>();
-
+    private final List<Long> boundariesOfCustomChromosomeX = new ArrayList<>();
+    private final List<Long> boundariesOfCustomChromosomeY = new ArrayList<>();
+    
     private static Pair<List<MotifAnchor>, List<MotifAnchor>> getAllRegionsFromSubChromosomes(
             final ChromosomeHandler handler, Chromosome chr) {
-
+        
         if (handler.isCustomChromosome(chr)) {
             final List<MotifAnchor> allRegions = new ArrayList<>();
             final List<MotifAnchor> translatedRegions = new ArrayList<>();
-
+            
             handler.getListOfRegionsInCustomChromosome(chr.getIndex()).processLists(
                     new juicebox.data.feature.FeatureFunction<MotifAnchor>() {
                         @Override
@@ -53,10 +53,10 @@ public class CustomMZDRegionHandler {
                         }
                     });
             Collections.sort(allRegions);
-
-            int previousEnd = 0;
+            
+            long previousEnd = 0;
             for (MotifAnchor anchor : allRegions) {
-                int currentEnd = previousEnd + anchor.getWidth();
+                long currentEnd = previousEnd + anchor.getWidth();
                 MotifAnchor anchor2 = new MotifAnchor(chr.getName(), previousEnd, currentEnd);
                 translatedRegions.add(anchor2);
                 previousEnd = currentEnd + ChromosomeHandler.CUSTOM_CHROMOSOME_BUFFER;
@@ -67,8 +67,8 @@ public class CustomMZDRegionHandler {
             // just a standard chromosome
             final List<MotifAnchor> allRegions = new ArrayList<>();
             final List<MotifAnchor> translatedRegions = new ArrayList<>();
-            allRegions.add(new MotifAnchor(chr.getName(), 0, chr.getLength()));
-            translatedRegions.add(new MotifAnchor(chr.getName(), 0, chr.getLength()));
+            allRegions.add(new MotifAnchor(chr.getName(), 0, (int) chr.getLength()));
+            translatedRegions.add(new MotifAnchor(chr.getName(), 0, (int) chr.getLength()));
             return new Pair<>(allRegions, translatedRegions);
         }
     }
@@ -88,19 +88,19 @@ public class CustomMZDRegionHandler {
             boundariesOfCustomChromosomeY.addAll(boundariesOfCustomChromosomeX);
         }
     }
-
-    public List<Integer> getBoundariesOfCustomChromosomeX() {
+    
+    public List<Long> getBoundariesOfCustomChromosomeX() {
         return boundariesOfCustomChromosomeX;
     }
-
-    public List<Integer> getBoundariesOfCustomChromosomeY() {
+    
+    public List<Long> getBoundariesOfCustomChromosomeY() {
         return boundariesOfCustomChromosomeY;
     }
-
-    private void populateRegions(Chromosome chr, ChromosomeHandler handler, List<Integer> boundaries, HiCZoom zoom) {
+    
+    private void populateRegions(Chromosome chr, ChromosomeHandler handler, List<Long> boundaries, HiCZoom zoom) {
         String name = chr.getName();
         Pair<List<MotifAnchor>, List<MotifAnchor>> allRegionsInfo = getAllRegionsFromSubChromosomes(handler, chr);
-
+        
         if (allRegionsInfo != null) {
             allRegionsForChr.put(name, allRegionsInfo);
             List<MotifAnchor> translatedRegions = allRegionsInfo.getSecond();
@@ -109,14 +109,14 @@ public class CustomMZDRegionHandler {
             }
         }
     }
-
-    public List<Pair<MotifAnchor, MotifAnchor>> getIntersectingFeatures(String name, int gx1, int gx2) {
-
+    
+    public List<Pair<MotifAnchor, MotifAnchor>> getIntersectingFeatures(String name, long gx1, long gx2) {
+        
         int idx1 = OneDimSearchUtils.indexedBinaryNearestSearch(
                 allRegionsForChr.get(name).getSecond(), new MotifAnchor(name, gx1, gx1), true);
         int idx2 = OneDimSearchUtils.indexedBinaryNearestSearch(
                 allRegionsForChr.get(name).getSecond(), new MotifAnchor(name, gx2, gx2), false);
-
+        
         final List<Pair<MotifAnchor, MotifAnchor>> foundFeatures = new ArrayList<>();
         for (int i = idx1; i <= idx2; i++) {
             foundFeatures.add(new Pair<>(
